requirements.txt
groq>=0.4.1
torch>=1.9.0
numpy>=1.21.0
Pillow>=8.0.0
requests>=2.25.0
